+++ tutorial/react/getting-started/storage/initial/App.tsx
@@ -7,9 +7,9 @@ export default function App() {
   const roomId = "{% ROOM_ID %}";
 
   return (
-    <RoomProvider id={roomId} initialPresence={{}}>
+    <RoomProvider id={roomId}>
       <ClientSideSuspense fallback={<div>Loading…</div>}>
-        {() => <Room />}
+        <Room />
       </ClientSideSuspense>
     </RoomProvider>
   );