+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## v2.0.4 (not released yet)
 
+### All packages
+
+- Improve TS error messages and error locations if custom `UserMeta` or
+  `ActivitiesData` types do not match their requirements
+
 ### `@liveblocks/client`
 
 - Add missing type export for `CommentReaction`
@@ -815,8 +820,7 @@ leave();
 ## v1.3.0
 
 This release marks the initial release of
-[Liveblocks Comments](https://liveblocks.io/comments), which is currently in
-private beta.
+[Liveblocks Comments](https://liveblocks.io/comments) (private beta).
 
 ### `@liveblocks/client`